+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Bugfix: Prevent concurrent accesses of eval event database from raising lock errors.
 - Inspect View: Properly wrap log configuration values in evaluation header.
+- Inspect View: Support for displaying and navigating directories of evaluation logs.
 
 ## v0.3.106 (21 June 2025)